[FEAT] gradle 종속성 추가, 파일 생성 위치바꿈
