+++ app/build.gradle.kts
@@ -48,10 +48,25 @@ dependencies {
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
+    implementation("androidx.activity:activity-ktx:1.8.2")
+    //ViewModel
+    implementation("androidx.lifecycle:lifecycle-livedata-ktx:2.7.0")
+    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:2.7.0")
     //Fragment
     implementation ("androidx.fragment:fragment-ktx:1.6.2")
+    //BottomNavigation
+    implementation("androidx.navigation:navigation-fragment-ktx:2.7.7")
+    implementation("androidx.navigation:navigation-ui-ktx:2.7.7")
     //RecordFragment Charts
     implementation ("com.diogobernardino:williamchart:3.10.1")
     //Naver Map
     implementation("com.naver.maps:map-sdk:3.17.0")
+    //Retrofit
+    implementation("com.google.code.gson:gson:2.10.1")
+    implementation("com.squareup.retrofit2:retrofit:2.9.0")
+    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
+    implementation("com.squareup.okhttp3:okhttp:4.10.0")
+    implementation("com.squareup.okhttp3:logging-interceptor:4.10.0")
+    //Glide
+    implementation("com.github.bumptech.glide:glide:4.16.0")
 }
