+++ app/src/main/java/com/twoday/todaytrip/ui/place_detail/PlaceDetailFragment.kt
@@ -1,0 +1,37 @@
+package com.twoday.todaytrip.ui.place_detail
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.twoday.todaytrip.R
+import com.twoday.todaytrip.databinding.FragmentPlaceDetailBinding
+
+class PlaceDetailFragment : Fragment() {
+
+    private var _binding: FragmentPlaceDetailBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        _binding = FragmentPlaceDetailBinding.inflate(layoutInflater)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+
+        return binding.root
+    }
+
+    companion object {
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            PlaceDetailFragment().apply {
+//                arguments = Bundle().apply {
+//                }
+            }
+    }
+}
