+++ app/src/main/java/com/twoday/todaytrip/ui/place_detail/PlaceDetailPhotoAdapter.kt
@@ -1,0 +1,4 @@
+package com.twoday.todaytrip.ui.place_detail
+
+class PlaceDetailPhotoAdapter {
+}
