+++ app/src/main/java/com/twoday/todaytrip/ui/place_map/PlaceMapActivity.kt
@@ -1,4 +1,4 @@
-package com.twoday.todaytrip.ui
+package com.twoday.todaytrip.ui.place_map
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -11,10 +11,14 @@ class PlaceMapActivity : AppCompatActivity() {
     private var _binding: ActivityPlaceMapBinding? = null
     private val binding get() = _binding!!
 
+    private lateinit var adapter: PlaceMapAdapter
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         _binding = ActivityPlaceMapBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
     }
+
+
 }
