+++ app/src/main/res/layout/fragment_place_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PlaceDetailFragment">
+    tools:context=".ui.place_detail.PlaceDetailFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/vp_place_detail_pic"
